train.py splitted for better organization
